+++ appFrontend/package.json
@@ -68,6 +68,7 @@
     "@react-native/metro-config": "0.73.5",
     "@react-native/typescript-config": "0.73.1",
     "@testing-library/react-native": "^12.3.0",
+    "@trivago/prettier-plugin-sort-imports": "^6.0.2",
     "@tsconfig/react-native": "^3.0.0",
     "@types/lodash": "^4.14.200",
     "@types/react": "^18.2.6",
