+++ appFrontend/src/screens/searchStationScreen/index.tsx
@@ -1,9 +1,4 @@
-import { useAddRecentSearch, useGetSearchHistory, useSearchStationName } from '@/global/apis/hooks';
-import { COLOR } from '@/global/constants';
-import { FontText } from '@/global/ui';
-import { displayToOrigin } from '@/global/utils/subwayLine';
-import { useAppDispatch, useAppSelect } from '@/store';
-import { getSeletedStation } from '@/store/modules/stationSearchModule';
+import { SafeAreaView } from 'react-native-safe-area-context';
 import { useState } from 'react';
 import {
   KeyboardAvoidingView,
@@ -13,18 +8,23 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native';
+import { getSeletedStation } from '@/store/modules/stationSearchModule';
+import { OriginLineName } from '@/global/apis/entity';
+import { useAddRecentSearch, useGetSearchHistory, useSearchStationName } from '@/global/apis/hooks';
+import { COLOR } from '@/global/constants';
+import { FontText } from '@/global/ui';
 import { Input } from '@/global/ui';
+import { displayToOrigin } from '@/global/utils';
 import { useHomeNavigation } from '@/navigation/HomeNavigation';
+import { trackMapSearchArrivalChoice, trackMapSearchDepartureChoice } from '@/analytics/map.events';
 import {
   IconArrowLeftSharp,
   IconClock,
+  IconCrossCircle,
   IconLocationPin,
   IconNoResult,
-  IconCrossCircle,
 } from '@/assets/icons';
-import { OriginLineName } from '@/global/apis/entity';
-import { trackMapSearchArrivalChoice, trackMapSearchDepartureChoice } from '@/analytics/map.events';
-import { SafeAreaView } from 'react-native-safe-area-context';
+import { useAppDispatch, useAppSelect } from '@/store';
 
 const SearchStationScreen = () => {
   const navigation = useHomeNavigation();
